src/main/webapp/app/entities/action/route/action-routing-resolve.service.spec.ts
jest.mock('@angular/router');

import { TestBed } from '@angular/core/testing';
import { HttpResponse } from '@angular/common/http';
import { HttpClientTestingModule } from '@angular/common/http/testing';
import { ActivatedRouteSnapshot, Router } from '@angular/router';
import { of } from 'rxjs';

import { IAction, Action } from '../action.model';
import { ActionService } from '../service/action.service';

import { ActionRoutingResolveService } from './action-routing-resolve.service';

describe('Service Tests', () => {
  describe('Action routing resolve service', () => {
    let mockRouter: Router;
    let mockActivatedRouteSnapshot: ActivatedRouteSnapshot;
    let routingResolveService: ActionRoutingResolveService;
    let service: ActionService;
    let resultAction: IAction | undefined;

    beforeEach(() => {
      TestBed.configureTestingModule({
        imports: [HttpClientTestingModule],
        providers: [Router, ActivatedRouteSnapshot],
      });
      mockRouter = TestBed.inject(Router);
      mockActivatedRouteSnapshot = TestBed.inject(ActivatedRouteSnapshot);
      routingResolveService = TestBed.inject(ActionRoutingResolveService);
      service = TestBed.inject(ActionService);
      resultAction = undefined;
    });

    describe('resolve', () => {
      it('should return IAction returned by find', () => {
        // GIVEN
        service.find = jest.fn(id => of(new HttpResponse({ body: { id } })));
        mockActivatedRouteSnapshot.params = { id: 123 };

        // WHEN
        routingResolveService.resolve(mockActivatedRouteSnapshot).subscribe(result => {
          resultAction = result;
        });

        // THEN
        expect(service.find).toBeCalledWith(123);
        expect(resultAction).toEqual({ id: 123 });
      });

      it('should return new IAction if id is not provided', () => {
        // GIVEN
        service.find = jest.fn();
        mockActivatedRouteSnapshot.params = {};

        // WHEN
        routingResolveService.resolve(mockActivatedRouteSnapshot).subscribe(result => {
          resultAction = result;
        });

        // THEN
        expect(service.find).not.toBeCalled();
        expect(resultAction).toEqual(new Action());
      });

      it('should route to 404 page if data not found in server', () => {
        // GIVEN
        jest.spyOn(service, 'find').mockReturnValue(of(new HttpResponse({ body: null as unknown as Action })));
        mockActivatedRouteSnapshot.params = { id: 123 };

        // WHEN
        routingResolveService.resolve(mockActivatedRouteSnapshot).subscribe(result => {
          resultAction = result;
        });

        // THEN
        expect(service.find).toBeCalledWith(123);
        expect(resultAction).toEqual(undefined);
        expect(mockRouter.navigate).toHaveBeenCalledWith(['404']);
      });
    });
  });
});
